Lion Bot V0.2
After my second session of programming I have bugfixed the initial bot and also added a lot more features. Firstly, there is randomisation within the move selection so that the bot doesn't do the same thing every time. Secondly, I've added Castling and it's somewhat forced. Third, I've removed the random King wanders in the middlegame and the King will not move until the Endgame begins, which begins when there are 6 total pieces or fewer remaining on the board - also a new implementation. The King will then activate in the Endgame and his moves will be added into the random pool. Finally, Pawn Promotions are now forced.